Modified Post Migration.

+++ database/migrations/2019_07_21_065745_create_posts_table.php
@@ -15,7 +15,7 @@ class CreatePostsTable extends Migration
     {
         Schema::create('posts', function (Blueprint $table) {
         	$table->bigIncrements('id');
-			$table ->integer('author_id')->unsigned()->default(0);
+			$table ->bigInteger('author_id')->unsigned()->default(0);
 			$table->foreign('author_id')->references('id')->on('users')->onDelete('cascade');
 			$table->string('title');
 			$table->text('content');
